Streamline PeerInfo struct
- do not require custom getters
- move instantiation logic into rust-only impl method
- improve formatting of JS logging
